+++ src/Collisions/Collisions.h
@@ -16,7 +16,7 @@ class Collisions
 
 public:
     //! Constructor for Collisions between two species
-    Collisions( Patch* patch, unsigned int n_collisions, std::vector<unsigned int>,std::vector<unsigned int>, double coulomb_log, bool intra_collisions,int debug_every,int Z,bool ionizing,int nDim,double);
+    Collisions( Patch* patch, unsigned int n_collisions, std::vector<unsigned int>,std::vector<unsigned int>, double coulomb_log, bool intra_collisions,int debug_every,int Z,bool ionizing,int nDim,double,std::string);
     //! Cloning Constructor
     Collisions(Collisions*, int);
     //! destructor
@@ -26,6 +26,23 @@ public:
     static std::vector<Collisions*> create(Params&, Patch*, std::vector<Species*>&);
     //! Method that clones a vector of Collisions objects
     static std::vector<Collisions*> clone(std::vector<Collisions*>, Params&);
+    
+    //! Method to calculate the Debye length in each cluster
+    static void calculate_debye_length(Params&, Patch*);
+    
+    //! is true if any of the collisions objects need automatically-computed coulomb log
+    static bool debye_length_required;
+    
+    //! Method called in the main smilei loop to apply collisions at each timestep
+    void collide(Params&, Patch* ,int);
+    
+    //! Outputs the debug info if requested
+    static void debug(Params& params, int itime, unsigned int icoll, VectorPatch& vecPatches);
+    
+    //! Deflection angle calculation
+    static double cos_chi(double);
+    
+private:
     
     //! Identification number of the Collisions object
     int n_collisions;
@@ -42,25 +59,10 @@ public:
     //! Number of timesteps between each dump of collisions debugging
     int debug_every;
     
-    //! Method to calculate the Debye length in each cluster
-    static void calculate_debye_length(Params&, Patch*);
-    
-    //! is true if any of the collisions objects need automatically-computed coulomb log
-    static bool debye_length_required;
-    
-    //! Method called in the main smilei loop to apply collisions at each timestep
-    void collide(Params&, Patch* ,int);
-    
-    //! Outputs the debug info if requested
-    static void debug(Params& params, int itime, unsigned int icoll, VectorPatch& vecPatches);
-    
     //! CollisionalIonization object, created if ionization required
     CollisionalIonization * Ionization;
     
-private:
-    
-    static double cos_chi(double);
-    
+    //! Species atomic number, in case of ionization
     int atomic_number;
     
     //! Hdf5 file name