+++ src/Patch/Patch.h
@@ -95,8 +95,10 @@ public:
     
     //! init comm / sum densities
     virtual void initSumField( Field* field, int iDim ) = 0;
+    virtual void reallyinitSumField( Field* field, int iDim ) = 0;
     //! finalize comm / sum densities
     virtual void finalizeSumField( Field* field, int iDim ) = 0;
+    virtual void reallyfinalizeSumField( Field* field, int iDim ) = 0;
     
     //! init comm / exchange fields
     virtual void initExchange( Field* field ) = 0;
@@ -106,7 +108,7 @@ public:
     virtual void initExchange( Field* field, int iDim ) = 0;
     //! finalize comm / exchange fields in direction iDim only
     virtual void finalizeExchange( Field* field, int iDim ) = 0;
-    
+
     // Create MPI_Datatype to exchange fields
     virtual void createType( Params& params ) = 0;
     
@@ -214,6 +216,9 @@ protected:
     std::vector< std::vector<int> > neighbor_;
     //! Hilbert index of corners neighbors patch
     std::vector< std::vector<int> > corner_neighbor_; // Kept for Moving Windows
+
+    std::vector< std::vector<int> > send_tags_, recv_tags_;
+
     
     //! MPI rank of neighbors patch
     std::vector< std::vector<int> > MPI_neighbor_;