+++ src/Patch/Patch1D.cpp
@@ -57,6 +57,27 @@ void Patch1D::initStep2(Params& params)
     
 }
 
+void Patch1D::reallyinitSumField( Field* field, int iDim )
+{
+    if (field->MPIbuff.srequest.size()==0)
+        field->MPIbuff.allocate(1);
+
+    std::vector<unsigned int> n_elem = field->dims_;
+    std::vector<unsigned int> isDual = field->isDual_;
+    Field1D* f1D =  static_cast<Field1D*>(field);
+    
+    // Use a buffer per direction to exchange data before summing
+    // Size buffer is 2 oversize (1 inside & 1 outside of the current subdomain)
+    std::vector<unsigned int> oversize2 = oversize;
+    oversize2[0] *= 2;
+    oversize2[0] += 1 + f1D->isDual_[0];
+    
+    for (int iNeighbor=0 ; iNeighbor<nbNeighbors_ ; iNeighbor++) {
+        std::vector<unsigned int> tmp(nDim_fields_,0);
+        tmp[0] =    iDim  * n_elem[0] + (1-iDim) * oversize2[0];
+        buf[iDim][iNeighbor].allocateDims( tmp );
+    }
+}
 
 // ---------------------------------------------------------------------------------------------------------------------
 // Initialize current patch sum Fields communications through MPI for direction iDim
@@ -77,11 +98,11 @@ void Patch1D::initSumField( Field* field, int iDim )
     oversize2[0] *= 2;
     oversize2[0] += 1 + f1D->isDual_[0];
     
-    for (int iNeighbor=0 ; iNeighbor<nbNeighbors_ ; iNeighbor++) {
+    /*for (int iNeighbor=0 ; iNeighbor<nbNeighbors_ ; iNeighbor++) {
         std::vector<unsigned int> tmp(nDim_fields_,0);
         tmp[0] =    iDim  * n_elem[0] + (1-iDim) * oversize2[0];
         buf[iDim][iNeighbor].allocateDims( tmp );
-    }
+    }*/
      
     int istart, ix;
     /********************************************************************************/
@@ -95,7 +116,7 @@ void Patch1D::initSumField( Field* field, int iDim )
         if ( is_a_MPI_neighbor( iDim, iNeighbor ) ) {
             istart = iNeighbor * ( n_elem[iDim]- oversize2[iDim] ) + (1-iNeighbor) * ( 0 );
             ix = (1-iDim)*istart;
-            int tag = buildtag( hindex, iDim, iNeighbor );
+            int tag = send_tags_[iDim][iNeighbor];
             MPI_Isend( &(f1D->data_[ix]), 1, ntype, MPI_neighbor_[iDim][iNeighbor], tag, MPI_COMM_WORLD, &(f1D->MPIbuff.srequest[iDim][iNeighbor]) );
             //MPI_Send( &(f1D->data_[ix]), 1, ntype, MPI_neighbor_[iDim][iNeighbor], tag, MPI_COMM_WORLD );
             //MPI_Isend( &(f1D->data_[ix]), iDim  * n_elem[0] + (1-iDim) * oversize2[0], MPI_DOUBLE, MPI_neighbor_[iDim][iNeighbor], tag, MPI_COMM_WORLD, &(f1D->MPIbuff.srequest[iDim][iNeighbor]) );
@@ -103,7 +124,7 @@ void Patch1D::initSumField( Field* field, int iDim )
             
         if ( is_a_MPI_neighbor( iDim, (iNeighbor+1)%2 ) ) {
             int tmp_elem = (buf[iDim][(iNeighbor+1)%2]).dims_[0];
-            int tag = buildtag( neighbor_[iDim][(iNeighbor+1)%2], iDim, iNeighbor );
+            int tag = recv_tags_[iDim][iNeighbor];
             MPI_Irecv( &( (buf[iDim][(iNeighbor+1)%2]).data_[0] ), tmp_elem, MPI_DOUBLE, MPI_neighbor_[iDim][(iNeighbor+1)%2], tag, MPI_COMM_WORLD, &(f1D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]) );
             //MPI_Status stat;
             //MPI_Recv( &( (buf[iDim][(iNeighbor+1)%2]).data_[0] ), tmp_elem, MPI_DOUBLE, MPI_neighbor_[iDim][(iNeighbor+1)%2], tag, MPI_COMM_WORLD, &stat );
@@ -147,8 +168,21 @@ void Patch1D::finalizeSumField( Field* field, int iDim )
             MPI_Wait( &(f1D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]), &(rstat[iDim][(iNeighbor+1)%2]) );
         }
     }
+}
 
-
+void Patch1D::reallyfinalizeSumField( Field* field, int iDim )
+{
+    std::vector<unsigned int> n_elem = field->dims_;
+    std::vector<unsigned int> isDual = field->isDual_;
+    Field1D* f1D =  static_cast<Field1D*>(field);
+   
+    // Use a buffer per direction to exchange data before summing
+    // Size buffer is 2 oversize (1 inside & 1 outside of the current subdomain)
+    std::vector<unsigned int> oversize2 = oversize;
+    oversize2[0] *= 2;
+    oversize2[0] += 1 + f1D->isDual_[0];
+    
+    int istart;
     /********************************************************************************/
     // Sum data on each process, same operation on both side
     /********************************************************************************/
@@ -197,7 +231,7 @@ void Patch1D::initExchange( Field* field )
 
                 istart = iNeighbor * ( n_elem[iDim]- (2*oversize[iDim]+1+isDual[iDim]) ) + (1-iNeighbor) * ( oversize[iDim] + 1 + isDual[iDim] );
                 ix = (1-iDim)*istart;
-                int tag = buildtag( hindex, iDim, iNeighbor );
+                int tag = send_tags_[iDim][iNeighbor];
                 MPI_Isend( &(f1D->data_[ix]), 1, ntype, MPI_neighbor_[iDim][iNeighbor], tag, MPI_COMM_WORLD, &(f1D->MPIbuff.srequest[iDim][iNeighbor]) );
 
             } // END of Send
@@ -206,7 +240,7 @@ void Patch1D::initExchange( Field* field )
 
                 istart = ( (iNeighbor+1)%2 ) * ( n_elem[iDim] - 1 - (oversize[iDim]-1) ) + (1-(iNeighbor+1)%2) * ( 0 )  ;
                 ix = (1-iDim)*istart;
-                 int tag = buildtag( neighbor_[iDim][(iNeighbor+1)%2], iDim, iNeighbor );
+                int tag = recv_tags_[iDim][iNeighbor];
                 MPI_Irecv( &(f1D->data_[ix]), 1, ntype, MPI_neighbor_[iDim][(iNeighbor+1)%2], tag, MPI_COMM_WORLD, &(f1D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]));
 
             } // END of Recv
@@ -268,7 +302,7 @@ void Patch1D::initExchange( Field* field, int iDim )
 
             istart = iNeighbor * ( n_elem[iDim]- (2*oversize[iDim]+1+isDual[iDim]) ) + (1-iNeighbor) * ( oversize[iDim] + 1 + isDual[iDim] );
             ix = (1-iDim)*istart;
-            int tag = buildtag( hindex, iDim, iNeighbor );
+            int tag = send_tags_[iDim][iNeighbor];
             MPI_Isend( &(f1D->data_[ix]), 1, ntype, MPI_neighbor_[iDim][iNeighbor], tag, MPI_COMM_WORLD, &(f1D->MPIbuff.srequest[iDim][iNeighbor]) );
 
         } // END of Send
@@ -277,7 +311,7 @@ void Patch1D::initExchange( Field* field, int iDim )
 
             istart = ( (iNeighbor+1)%2 ) * ( n_elem[iDim] - 1 - (oversize[iDim]-1) ) + (1-(iNeighbor+1)%2) * ( 0 )  ;
             ix = (1-iDim)*istart;
-            int tag = buildtag( neighbor_[iDim][(iNeighbor+1)%2], iDim, iNeighbor );
+            int tag = recv_tags_[iDim][iNeighbor];
             MPI_Irecv( &(f1D->data_[ix]), 1, ntype, MPI_neighbor_[iDim][(iNeighbor+1)%2], tag, MPI_COMM_WORLD, &(f1D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]));
 
         } // END of Recv