+++ src/Patch/Patch2D.cpp
@@ -82,12 +82,7 @@ void Patch2D::initStep2(Params& params)
 
 }
 
-
-// ---------------------------------------------------------------------------------------------------------------------
-// Initialize current patch sum Fields communications through MPI in direction iDim
-// Intra-MPI process communications managed by memcpy in SyncVectorPatch::sum()
-// ---------------------------------------------------------------------------------------------------------------------
-void Patch2D::initSumField( Field* field, int iDim )
+void Patch2D::reallyinitSumField( Field* field, int iDim )
 {
     if (field->MPIbuff.srequest.size()==0)
         field->MPIbuff.allocate(2);
@@ -120,6 +115,45 @@ void Patch2D::initSumField( Field* field, int iDim )
             tmp[1] = 1;
         buf[iDim][iNeighbor].allocateDims( tmp );
     }
+}
+
+// ---------------------------------------------------------------------------------------------------------------------
+// Initialize current patch sum Fields communications through MPI in direction iDim
+// Intra-MPI process communications managed by memcpy in SyncVectorPatch::sum()
+// ---------------------------------------------------------------------------------------------------------------------
+void Patch2D::initSumField( Field* field, int iDim )
+{
+    if (field->MPIbuff.srequest.size()==0)
+        field->MPIbuff.allocate(2);
+
+    int patch_ndims_(2);
+    int patch_nbNeighbors_(2);
+    
+    
+    std::vector<unsigned int> n_elem = field->dims_;
+    std::vector<unsigned int> isDual = field->isDual_;
+    Field2D* f2D =  static_cast<Field2D*>(field);
+   
+    // Use a buffer per direction to exchange data before summing
+    //Field2D buf[patch_ndims_][patch_nbNeighbors_];
+    // Size buffer is 2 oversize (1 inside & 1 outside of the current subdomain)
+    std::vector<unsigned int> oversize2 = oversize;
+    oversize2[0] *= 2;
+    oversize2[0] += 1 + f2D->isDual_[0];
+    if (field->dims_.size()>1) {
+        oversize2[1] *= 2;
+        oversize2[1] += 1 + f2D->isDual_[1];
+    }
+    
+    /*for (int iNeighbor=0 ; iNeighbor<patch_nbNeighbors_ ; iNeighbor++) {
+        std::vector<unsigned int> tmp(patch_ndims_,0);
+        tmp[0] =    iDim  * n_elem[0] + (1-iDim) * oversize2[0];
+        if (field->dims_.size()>1)
+            tmp[1] = (1-iDim) * n_elem[1] +    iDim  * oversize2[1];
+        else 
+            tmp[1] = 1;
+        buf[iDim][iNeighbor].allocateDims( tmp );
+    }*/
      
     int istart, ix, iy;
     /********************************************************************************/
@@ -134,8 +168,7 @@ void Patch2D::initSumField( Field* field, int iDim )
             istart = iNeighbor * ( n_elem[iDim]- oversize2[iDim] ) + (1-iNeighbor) * ( 0 );
             ix = (1-iDim)*istart;
             iy =    iDim *istart;
-            //int tag = buildtag( 3, hindex, neighbor_[iDim][iNeighbor]);
-            int tag = buildtag( hindex, iDim, iNeighbor );
+            int tag = send_tags_[iDim][iNeighbor];
             //cout << hindex << " send to " << neighbor_[iDim][iNeighbor] << endl;
             //MPI_Isend( &(f2D->data_2D[ix][iy]), 1, ntype, 0, tag, MPI_COMM_SELF, &(f2D->MPIbuff.srequest[iDim][iNeighbor]) );
             MPI_Isend( &(f2D->data_2D[ix][iy]), 1, ntype, MPI_neighbor_[iDim][iNeighbor], tag, MPI_COMM_WORLD, &(f2D->MPIbuff.srequest[iDim][iNeighbor]) );
@@ -143,8 +176,7 @@ void Patch2D::initSumField( Field* field, int iDim )
             
         if ( is_a_MPI_neighbor( iDim, (iNeighbor+1)%2 ) ) {
             int tmp_elem = (buf[iDim][(iNeighbor+1)%2]).dims_[0]*(buf[iDim][(iNeighbor+1)%2]).dims_[1];
-            //int tag = buildtag( 3, neighbor_[iDim][(iNeighbor+1)%2], hindex);
-            int tag = buildtag( neighbor_[iDim][(iNeighbor+1)%2], iDim, iNeighbor );
+            int tag = recv_tags_[iDim][iNeighbor];
             //cout << hindex << " recv from " << neighbor_[iDim][(iNeighbor+1)%2] << " ; n_elements = " << tmp_elem << endl;
             //MPI_Irecv( &( (buf[iDim][(iNeighbor+1)%2]).data_2D[0][0] ), tmp_elem, MPI_DOUBLE, 0, tag, MPI_COMM_SELF, &(f2D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]) );
             MPI_Irecv( &( (buf[iDim][(iNeighbor+1)%2]).data_2D[0][0] ), tmp_elem, MPI_DOUBLE, MPI_neighbor_[iDim][(iNeighbor+1)%2], tag, MPI_COMM_WORLD, &(f2D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]) );
@@ -195,8 +227,26 @@ void Patch2D::finalizeSumField( Field* field, int iDim )
             MPI_Wait( &(f2D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]), &(rstat[iDim][(iNeighbor+1)%2]) );
         }
     }
+}
 
+void Patch2D::reallyfinalizeSumField( Field* field, int iDim )
+{
+    int patch_ndims_(2);
+    int patch_nbNeighbors_(2);
+    std::vector<unsigned int> n_elem = field->dims_;
+    std::vector<unsigned int> isDual = field->isDual_;
+    Field2D* f2D =  static_cast<Field2D*>(field);
+   
+    // Use a buffer per direction to exchange data before summing
+    //Field2D buf[patch_ndims_][patch_nbNeighbors_];
+    // Size buffer is 2 oversize (1 inside & 1 outside of the current subdomain)
+    std::vector<unsigned int> oversize2 = oversize;
+    oversize2[0] *= 2;
+    oversize2[0] += 1 + f2D->isDual_[0];
+    oversize2[1] *= 2;
+    oversize2[1] += 1 + f2D->isDual_[1];
 
+    int istart;
     /********************************************************************************/
     // Sum data on each process, same operation on both side
     /********************************************************************************/
@@ -251,8 +301,7 @@ void Patch2D::initExchange( Field* field )
                 istart = iNeighbor * ( n_elem[iDim]- (2*oversize[iDim]+1+isDual[iDim]) ) + (1-iNeighbor) * ( oversize[iDim] + 1 + isDual[iDim] );
                 ix = (1-iDim)*istart;
                 iy =    iDim *istart;
-                //int tag = buildtag( 4, hindex, neighbor_[iDim][iNeighbor]);
-                int tag = buildtag( hindex, iDim, iNeighbor );
+                int tag = send_tags_[iDim][iNeighbor];
                 //MPI_Isend( &(f2D->data_2D[ix][iy]), 1, ntype, 0, tag, MPI_COMM_SELF, &(f2D->MPIbuff.srequest[iDim][iNeighbor]) );
                 MPI_Isend( &(f2D->data_2D[ix][iy]), 1, ntype, MPI_neighbor_[iDim][iNeighbor], tag, MPI_COMM_WORLD, &(f2D->MPIbuff.srequest[iDim][iNeighbor]) );
 
@@ -263,8 +312,7 @@ void Patch2D::initExchange( Field* field )
                 istart = ( (iNeighbor+1)%2 ) * ( n_elem[iDim] - 1 - (oversize[iDim]-1) ) + (1-(iNeighbor+1)%2) * ( 0 )  ;
                 ix = (1-iDim)*istart;
                 iy =    iDim *istart;
-                 //int tag = buildtag( 4, neighbor_[iDim][(iNeighbor+1)%2], hindex);
-                 int tag = buildtag( neighbor_[iDim][(iNeighbor+1)%2], iDim, iNeighbor );
+                int tag = recv_tags_[iDim][iNeighbor];
                 //MPI_Irecv( &(f2D->data_2D[ix][iy]), 1, ntype, 0, tag, MPI_COMM_SELF, &(f2D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]));
                 MPI_Irecv( &(f2D->data_2D[ix][iy]), 1, ntype, MPI_neighbor_[iDim][(iNeighbor+1)%2], tag, MPI_COMM_WORLD, &(f2D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]));
 
@@ -330,7 +378,7 @@ void Patch2D::initExchange( Field* field, int iDim )
             istart = iNeighbor * ( n_elem[iDim]- (2*oversize[iDim]+1+isDual[iDim]) ) + (1-iNeighbor) * ( oversize[iDim] + 1 + isDual[iDim] );
             ix = (1-iDim)*istart;
             iy =    iDim *istart;
-            int tag = buildtag( hindex, iDim, iNeighbor );
+            int tag = send_tags_[iDim][iNeighbor];
             //MPI_Isend( &(f2D->data_2D[ix][iy]), 1, ntype, 0, tag, MPI_COMM_SELF, &(f2D->MPIbuff.srequest[iDim][iNeighbor]) );
             MPI_Isend( &(f2D->data_2D[ix][iy]), 1, ntype, MPI_neighbor_[iDim][iNeighbor], tag, MPI_COMM_WORLD, &(f2D->MPIbuff.srequest[iDim][iNeighbor]) );
 
@@ -341,7 +389,7 @@ void Patch2D::initExchange( Field* field, int iDim )
             istart = ( (iNeighbor+1)%2 ) * ( n_elem[iDim] - 1- (oversize[iDim]-1) ) + (1-(iNeighbor+1)%2) * ( 0 )  ;
             ix = (1-iDim)*istart;
             iy =    iDim *istart;
-            int tag = buildtag( neighbor_[iDim][(iNeighbor+1)%2], iDim, iNeighbor );
+            int tag = recv_tags_[iDim][iNeighbor];
             //MPI_Irecv( &(f2D->data_2D[ix][iy]), 1, ntype, 0, tag, MPI_COMM_SELF, &(f2D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]));
             MPI_Irecv( &(f2D->data_2D[ix][iy]), 1, ntype, MPI_neighbor_[iDim][(iNeighbor+1)%2], tag, MPI_COMM_WORLD, &(f2D->MPIbuff.rrequest[iDim][(iNeighbor+1)%2]));